python/msgpack/requirements.txt
msgpack-python==0.4.8
u-msgpack-python==2.4.1
